#10136: Search for Map CRS coordinates Description: - resolve a threshold in CRS coordinate in switch
